ROCK PAPER SISSORS Created function to determine who won the game based on the number of rounds player won.
A dictionary is used to track the round win counts for the user
and the computer. Within the who_wins_game function, the dictionary
int value for each key which is either user or computer is incremented
based on who wins the round.

At the end of the round plays, the game winner is determined by
comparing the count values for both the user and computer players.
The player with the highest round count value wins the entire game.
Otherwise the game is tied.