DOWNLOAD ORGANIZER
Implemented logging to log error and info messages to separate log
files. The log files are stored in a logs directory which is
automatically created if it doesnt exists.

Error and Info messsages are logged throughout the script, to help with
debuging.

Also added more exception handling.

Added a file processing timeout feature which moves to next file
for processing after a specified amount of time, so as not to
spend a long time downloading one file.