CALCULATOR
added the art.py file ffrom which the calculator welcom art is imported
from and displayed at start of the calculator program.

completed the get_operation() function to allow user to
choose a math operation. user is asked for a math operation
until a valid operation from the list provided is selected.