LEAP YEAR
Added an exit_program function to centralize the program capability.
Refactored the logic to determine if a year is leap, made it more
concise.
Refactored how to determine if entered year is a four digit number using
a range 1000 to 9999.
Removed outter try...except block from the get_year function, and handle
any other errors raised in the main functions try...except block.