TEMPERATURE CONVERTER
Created constants for standard error messages and inuts and isolated in
a separate module.

Created an Enum for the conversion types in a separate module.

Refactored the perform_conversion function to use a lookup dictionary
for the types of conversions each having a value that is a lambda
function with the formula for the type of conversion.